+++ backend/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "server",
+    "name": "backend",
     "version": "0.0.1",
     "description": "",
     "author": "",
