+++ backend/src/users/controller/user.controller.ts
@@ -1,5 +1,5 @@
 import {JwtService} from '@nestjs/jwt';
-import {UserService} from './../service/user.service';
+import {UserService} from '../service/user.service';
 import {Body, Controller, HttpStatus, Post, Res} from '@nestjs/common';
 import {User} from '../schema/user.schema';
