+++ backend/src/videos/service/video.service.ts
@@ -1,4 +1,4 @@
-import {Video, VideoDocument} from './../schema/video.schema';
+import {Video, VideoDocument} from '../schema/video.schema';
 import {Injectable, NotFoundException, ServiceUnavailableException} from '@nestjs/common';
 import {InjectModel} from '@nestjs/mongoose';
 import {Model} from 'mongoose';
